services/orchestrator/requirements.txt
fastapi
uvicorn[standard]
langchain
langgraph
langchain-google-genai
langchain-openai
python-frontmatter
python-dotenv
httpx